Fixed a regression where the clipped mesh property was not working correctly.
Pushed a minimal CI to hopefully ensure that silly regressions such as merge conflicts don't en up on main.

Finished up the first pass on the color and demo integrations.

This integrates the existing work as well as creates a shader at compile time so that there's no additional configuration for the egui node.
The color tests don't fully pass, this is likely due to a blend mode issue.
TODO: Include an FAQ to explain certain smaller issues related to color.

Minor cleanup to these changes along with and FAQ.

Co-authored-by: derivator <derivator@users.noreply.github.com>

Updated the `egui2godot.shader` to use the correct order of operations. This should resolve the text blurriness that was occuring in the integration.

Added a check to skip assigning a material to any canvas item that is printing a User texture. This resolves an issue where you have to use specific settings to be compatible with egui.

Final tweaks to ensure that the samples all have GUI input when they are started.

Removed errant Cargo.lock file that should have been deleted.